Make all Arizona counties in the America/Phoenix tz
In some of the AZ counties, Navajo reservations follow daylight savings time, which would put them in the Denver tz. However, it looks like a lot of the population for some of these counties is in the part of the county that follows Arizona time (e.g., Flagstaff is in Coconino County).